kubevirt: Refactor the provider to follow the proxmoxve structure
The proxmoxve provider has a well-structured design where the provider's
public interface is a ConfigDrive that handles mounting the config drive
device, while the cloud-init parsing logic is isolated in cloudinit.rs.
This parser expects a path to a cloud-init directory, which makes testing
the trait interfaces directly much easier.

This change prepares the kubevirt provider to add support for cloud-init
network data that will be parsed similarly to proxmoxve.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>